Fix various cases of undefined behaviour on shifts
Mainly due to annoying C type promotion which by default
promotes unsigned char to signed int.

Bugs found by libfuzzer / undefined behaviour sanitizer while
testing OSS-Fuzz.

+++ cram/rANS_static.c
@@ -221,8 +221,10 @@ unsigned char *rans_uncompress_O0(unsigned char *in, unsigned int in_size,
     if (*in++ != 0) // Order-0 check
         return NULL;
 
-    in_sz  = ((in[0])<<0) | ((in[1])<<8) | ((in[2])<<16) | ((in[3])<<24);
-    out_sz = ((in[4])<<0) | ((in[5])<<8) | ((in[6])<<16) | ((in[7])<<24);
+    in_sz  = ((((uint32_t) in[0])<<0) | (((uint32_t) in[1])<<8) |
+              (((uint32_t) in[2])<<16) | (((uint32_t) in[3])<<24));
+    out_sz = ((((uint32_t) in[4])<<0) | (((uint32_t) in[5])<<8) |
+              (((uint32_t) in[6])<<16) | (((uint32_t) in[7])<<24));
     if (in_sz != in_size-9)
         return NULL;
 
@@ -586,8 +588,10 @@ unsigned char *rans_uncompress_O1(unsigned char *in, unsigned int in_size,
     if (*in++ != 1) // Order-1 check
         return NULL;
 
-    in_sz  = ((in[0])<<0) | ((in[1])<<8) | ((in[2])<<16) | ((in[3])<<24);
-    out_sz = ((in[4])<<0) | ((in[5])<<8) | ((in[6])<<16) | ((in[7])<<24);
+    in_sz  = ((((uint32_t) in[0])<<0) | (((uint32_t) in[1])<<8) |
+              (((uint32_t) in[2])<<16) | (((uint32_t) in[3])<<24));
+    out_sz = ((((uint32_t) in[4])<<0) | (((uint32_t) in[5])<<8) |
+              (((uint32_t) in[6])<<16) | (((uint32_t) in[7])<<24));
     if (in_sz != in_size-9)
         return NULL;
